Initial Commit
Includes Power Automate solution, instructions, Excel file template. 

This automation uses MS Graph to enumerate all SharePoint columns from all SharePoint lists across all SharePoint sites in a tenant. The point of this automation is to catalog this data. Very often SharePoint can be overused in ways that are not fully realized. The overuse contributes to multiple versions of the truth, storage of sensitive enterprise and personal data that may not be appropriate for SharePoint and duplication of data within SharePoint (user identity data is often part of each SharePoint list). The output of this automation should help to illuminate the extent to which SharePoint has been used and provide a bit of evidence to manage enterprise data.